requirements.txt
Django==1.7.6
Markdown==2.6.1
PAM==0.4.2
Pillow==2.3.0
SQLAlchemy==0.9.9
Twisted-Core==13.2.0
Twisted-Web==13.2.0
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
csvkit==0.9.0
dbf==0.94.003
debtagshw==0.1
defer==1.0.6
dirspec==13.10
django-filter==0.9.2
djangorestframework==3.1.0
duplicity==0.6.23
html5lib==0.999
httplib2==0.8
jdcal==1.0
lockfile==0.8
lxml==3.3.3
mysql-connector-python==1.1.6
mysql-utilities==1.3.5
oauthlib==0.6.1
oneconf==0.3.7
openpyxl==2.2.0
paramiko==1.10.1
pexpect==3.1
piston-mini-client==0.7.5
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pyserial==2.6
pysmbc==1.0.14.1
pysqlite==2.6.3
python-apt==0.9.3.5ubuntu1
python-dateutil==2.2
python-debian==0.1.21-nmu2ubuntu2
pyxdg==0.25
reportlab==3.0
requests==2.2.1
sessioninstaller==0.0.0
six==1.9.0
software-center-aptd-plugins==0.0.0
system-service==0.1.6
unity-lens-photos==1.0
urllib3==1.7.1
wsgiref==0.1.2
xdiagnose==3.6.3build2
xlrd==0.9.3
xmlutils==1.1
zope.interface==4.0.5
